arreglados los argumentos, test semi funcionando

+++ Userland/SampleCodeModule/include/test_util.h
@@ -10,6 +10,6 @@ int64_t satoi(char *str);
 void bussy_wait(uint64_t n);
 void endless_loop();
 void endless_loop_print(uint64_t wait);
-int64_t test_processes(uint64_t argc, char *argv[]);
+int64_t test_processes(int argc, char *argv[]);
 
 #endif
